requirements.txt
Django
djangorestframework
django-crontab
